backend/requirements.txt
django-filter==25.1
django==4.2.20
djangorestframework==3.15.2
djoser==2.3.1
gunicorn==23.0.0
pillow==11.1.0
psycopg2-binary==2.9.10
reportlab==4.3.1
